tests/py: Enable umockdev based tests
We re-exec the test case in pytest_configure if the LD_PRELOAD doesn't
contain libumockdev-preload.so. Even if the system doesn't have it, it
will show up in the environment and we can continue.

It also sets the UMOCKDEV_DIR env for the portal frontend. In theory
dbusmock should already do this for us but it doesn't seem to be
working.

Finally, it makes sure that ASAN keeps working even when LD_PRELOAD is
being used which is always the case because we require umockdev.