Allow host apps to register themselves
For host apps, we currently require applications to run inside a cgroup
with a name that allows xdg-desktop-portal to derive an application ID
from it. This does not work reliably, and while the situation can
improve, it's unlikely to ever be fool proof, as it in theory requires
every potential piece of software that launches other applications to
know how to launch it in a compatible way.

Cgroups is also a Linux specific API, which is problematic from a *nix
wide portability perspective.

To address this, introduce a host-only portal API that allows
applications to register their D-Bus connections, associating them with
an application ID. This moves back the burden of associating an
application with an application ID back to the application itself,
instead of the application launcher.

Applications calling this overrides any applicion ID derived from the
name fo the cgroup the application runs inside.