tests: Run all tests once for host, flatpak and snap when possible
By parametrizing the default xdp_app_info fixture, we can run all tests
once with a host, flatpak and snap XdpAppInfo and thus test all the
different code paths that can be hit.

This has historically been a blind spot when it comes to testing.

It's still possible to overwrite the fixture when a certain test only
works for host apps (e.g. the Registry one), or other subsets.

It's also possible to specify more detailed XdpAppInfo information.

Tests can query the app info kind and adjust their expected behavior
accordingly which is required in some cases.

The big drawback is that the number of test cases is increased by almost
3x, and so is the runtime.