request-future: Add a new Request implementation which uses futures
This also allows us to change the design a bit: requests get created by
the dbus method handler functions and are owned. When the ref-count
drops to zero, the Request gets closed and unexported. This allows using
an autoptr in the handler for ownership, without ever having to owne it
somewhere else.